fix: resolve bug with bottom bar interaction
- Added `zIndex: 1` in the `Animated.View` to prevent it from blocking after the animation ends.
- Moved the animation to the `TabBar` component for better code encapsulation.
- Improved related documentation for better understanding of the code.
